Added client- and server side validation tests for callback URLs.
